refactor: awaiters -> read queue
